Dear all.
This is the second major change of the script, but probably the
last one from my side so far. I'm sorry but I had lots of fun
and many ideas during the last days.
In future I will perform only small changes from time to time.
Promised.

To my opinion the script is ready to go. It is a stable version
with lots of nice features. It is completely sufficient to load 
atoms and sticks from PDB files and to modify the atoms afterwards.

Note that the major functionalities and the structure of the script 
have not been changed. Please, read the comments at the end of 
this text!


Changes
=======

1. The file operators have been removed in the panel and an
entry in the menu 'File -> Import' is created after the call
of the script. Its name is 'PDB (.pdb)' what I have adapted 
to what one can see in the Import sub-menu. 

2. The data file: Well, I organized the script such that 
the data file is automatically loaded. The condition is that the
data file is in the same directory as the python script.
Remember that the data file is important: It includes all colors 
and radii of the atoms, which are not included in PDB files.
So far, you need to manually load the script into Blender. Depending
on some decisions (see below) this can be easily changed.
TNa changes in data file.

3. I considered also this 'last active object' issue, Campbell is 
mentioning on the page here: 
http://codereview.appspot.com/5369091/diff/1//dsk/data/src/blender/blender/release/scripts/addons/io_pdb_atomic.py
I changed all and hope that it is fine now.

4. I re-arranged a bit the buttons, in- and outputs. All buttons,
inputs and outputs have now an identifier in Python, which contains 
the string 'atom_pdb'. In future, other scripts refer to all this. 

5. I added some buttons for the scaling of the atom radii. The
meaning is somewhat self-explanatory.

6. I added also some comments in the code. To my opinion it is useful 
for others for understanding the script, but also for newcomers who 
want to understand the meaning of some specific Blender commands.

7. I modified also the camera code. Euler math is quite nice and
finally I could do some mathematics. ... The camera is now in
a better orientation.

8. I removed some bugs and optimized the code.


Comments
========

1. I don't know if I meet the standards of Blender although I have 
given my very best. If there is something not convenient then help 
me, please. Thanks. 

2. To my opinion, all buttons and inputs/outputs in the panel are 
useful for well preparing the objects. It would be nice if they could
stay. I do use them when I charge molecules or crystal structures.
Note that I plan a second, bigger project where I need all this stuff.

3. There needs to be a decision done: In which tab shall the panel 
go? Physics? In fact, a separate tab only for this script would be quite perfect. 
Is this possible? - I couldn't find any code, examples, etc ... .

4. A second decision needs to be done, too: Where shall the script go?
Does it become a part of Blender or is it an add-on or is it simply a 
python script?
Depending on all this, some changes need to be done in the script: If it
is in some directories like .blender/scripts/startup one needs to
change the way for determing the data file. <= Tell me and I change the 
code, it's easy.

5. Which version shall the script obtain? It has 1.1 now, which
is my old way of counting. Normally, it would get version 2.0 because of 
all the major changes I have done. What do you think?

6. Shall we rename the file name (only io_import_pdb.py)? Well, the name 
'PDB - Atomic Blender' can stay in the Panel, if you agree (would be nice). 
In some time there will be 'PDB - Atomic Blender Movie'. ;-)

7. Last thing: Somebody asked me if one could also read files from
some quantum chemical simulation programs like VASP (http://cms.mpi.univie.ac.at/vasp/),
SIESTA (http://www.icmab.es/siesta-joomla/) etc. I can deal with this, too, 
but I need some time for this. Is there some demand from your side, 
do you have an interest?

8. If you have any other fancy ideas or changes to do, feel free to
modify the script!


So far ... I have no clue what follows now. So what is coming next? 
What are your comments? 


All the best and cheers,

Blendphys.




     


git-svn-id: https://svn.blender.org/svnroot/bf-extensions/contrib/py/scripts/addons@2627 96b784f8-4ae3-dd11-aff7-0030488c597c