Dear all.
Yet again an overall polishing of my script. I'm sorry but there are lots
of changes in the code and treatment of the structures. However, the script is
much more convenient now.

Basic changes
=============

1. I removed my first loading procedure, which has taken simply too much time 
for loading a many-atom scenario. Only the dupliverts technique is used now.

2. Ideasman told me that care must be taken when changing the properties of
objects. For instance, the objects in other layers or scenes should not be 
modified. Following the advices, I was dealing with this issue.
   1. One can now change the radii either of all objects, which are activated 
      (marked) in the current scene and layer.
   2. Or one can change the radii of all atoms in one scene and in all layers
      which are active. This is useful since one wishes to do that from time 
      to time.
      
3. The type of radii can be changed: pre-defined radii (pre-defined in a custom
data file), atomic radii and van der Waals radii. The pre-defined radii in the
script are in fact atomic radii.

4. An issue with this dupliverts technique is that a single representative ball is
reproduced on a mesh, which reflects the structure of one type of atom (sub 
atomic structure) - a single atom cannot be really independently treated. To
extract an atom that has been chosen via the vertex in the 'edit mode' by the 
user, a button has been introduced into the panel. The button appears as soon 
as one is in the 'edit mode'. Pushing the button executes code, which separates 
the marked atom from the mesh or dupliverts structure such that the atom is a 
separate ball. This ball is completely independent, and its properties can be 
changed accordingly independently (shape, color, etc.).

5. There were some bugs in the code which reads PDB files. Also, the element 
list was 'strange' at some places (some elements were missing and all types of 
radii had the values of the atomic radii). All bugs could be removed.

6. I forgot to mention that a part of the buttons are now in the file dialog 
window. This is more convenient.




I consider the script as finished. I give it the version 0.9 since some changes
will be done in future during refereeing the manuscript. Here are some PDB 
files which you can try out. Have fun to try out all the stuff:

http://qsad.bu.edu/data/pdbfiles/
http://www.pdb.org/pdb/home/home.do
http://development.root-1.de/Atomic_Blender/

At the end I recognize that it is not much. I have spent most of the time with
the Blender specific Python code - it is not easy and not at all fast to get 
what one wishes! So far, I somewhat familiar with Blender python and changes
with respect to PDB etc. are easy.




Isssues
=======

- Measuring distance between two active objects in the object mode works. 
However, the same doesn't work with two vertices marked in the edit mode 
so far. The latter is due to the fact that I cannot determine the active 
vertices. ... otherwise you tell me how ...

- Separating an atom from a dupliverts structure works. However, it does not 
work if one wants to do the same with more than 1 activated atom. It is again
due to the issue of active vertices, which cannot be determined.

- I don't really like the place where the panel appears (VIEW_3D, TOOL_PROPS).
I want to have my own tab! :-)
Why? Right now, there are many other panels above my on always pushing my nice 
'Atomic Blender' panel to the bottom. Hmmm ... but I have no idea where to put
it elsewhere.


To be done in future
====================

- I would like to make more order in the structures, which appear in the 
outliner. For instance, sub-atomic structures (hydrogens, carbons, etc) should 
be put into a single structure called 'molecule' or 'crystal' or 'structure'. 
This can obviously be done via parenting.

- Double or triple bonds are not represented by sticks so far. However, this is 
needed. <= Very easy to do and is next to get included 

- I would like to choose also the ionic radii. But this is not easy to integrate
into the script: Note that not all atoms have same charge states and that a
single atom can have several charge states.

- Theorists use a file type which is called .xyz. This file can include many
configurations of a molecule, surface or complex atomic structure. For instance, 
such files are used to represent structures composed by atoms which move in 
time. <= My next project.
Integrating this should be easy since I have learned a lot about Blender's 
python interface. <= Is next to be done.


Ackknowledgements: 
==================
Thanks to ideasman, meta_androcto, truman, kilon, dairin0d, PKHG, Valte, etc. 
for helping me with the code. I feel sorry for the many 'stupid' questions I 
have been posing. (and don't fell on the save side now since I stop working on
the script: I'm going to annoy you also in the next future. ;-))


So, I wait now for feedback. What do you all think about the manuscript?


Cheers all,

Blendphys.






git-svn-id: https://svn.blender.org/svnroot/bf-extensions/contrib/py/scripts/addons@2670 96b784f8-4ae3-dd11-aff7-0030488c597c