Bundled Material Library Maintenance
Material previews are no longer needed for my add-on to work with the current Blender, so I have gone and removed them.
As well, I have made some changes to OSL materials; some still need confirmation that they are indeed CC-0, and I have hidden them in the index.xml file with comment brackets.
All non-hidden OSL materials now have their OSL script nodes referencing external .osl files, stored in the scripts/ folder.

This was possible with version 0.5 of my Online/Offline material library add-on, but it was not employed.
However, with version 0.6 came the ability to view not only a library's materials, but the image textures, OSL scripts, and nodegroups it uses, so those external OSL scripts are now also displayed in the library.



git-svn-id: https://svn.blender.org/svnroot/bf-extensions/contrib/py/scripts/addons@4264 96b784f8-4ae3-dd11-aff7-0030488c597c