Initial commit.
Identical to the 0.0.8 release on GitHub https://github.com/varkenvarken/spacetree/tree/master/release

add_mesh_space_tree is a tree generation add-on based on the space colonization algorithm by A. Runions.
It is aimed as an alternative (not a replacement) to Sapling.

currently full docs are at http://blenderthings.blogspot.nl/p/the-spacetree-blender-add-on.html but these will be moved to the BlenderWiki ar
http://wiki.blender.org/index.php/Extensions:2.6/Py/Scripts/Add_Mesh/Add_Space_Tree
 
Next steps: transferring documentation to wiki, updating bl_info, and pep8 compliance. 

git-svn-id: https://svn.blender.org/svnroot/bf-extensions/contrib/py/scripts/addons@4637 96b784f8-4ae3-dd11-aff7-0030488c597c