Experimental keymap: further tweaks to subdivision level shifting.
The minimum subdivision level is now always zero, and the maximum is
the render subdivision level.  This seems sensible, as it's unlikely
the user will want to shift to a subdivision level higher than the
render level.  It also provides flexibility, so that the user can\
effectively set their own maximum for a given model.

Also, misc tweaks to other parts of the keymap, to make things
function a little more smoothly.


git-svn-id: https://svn.blender.org/svnroot/bf-extensions/contrib/py/scripts/addons@4743 96b784f8-4ae3-dd11-aff7-0030488c597c